baselines: 3 new arms + per-layer-KV read path; data/objective fixes; graph-memory design docs
BASELINES (all faithful to their papers — adversarial audit, 0 unintentional misrepresentations;
matched M=96 + ~7M trainable params, 4.8% spread):
- NEW gisting (per-layer gist-KV), memoryllm (per-layer pool + random-drop), titans (deep-MLP
  test-time autograd memory, MAC prepend) encoders + registry + capacity presets.
- Beacon REFIT from last-hidden-prepend to its native per-layer KV read.
- AutoCompressor refactored to faithful summary-accumulation (was secretly RMT).
- Shared per-layer-KV read path: decoder.build_prefix_cache + model._prefix_kv_forward, wired into
  both the MAE and QA read paths (native read for beacon/gisting/memoryllm); prepend path untouched.
- _OBJ_OK whitelist now admits the aux-loss-free new arms under behavioral_kl.

DATA / OBJECTIVE fixes:
- behavioral_kl teacher: disable_lora() so it is a genuinely frozen, stationary reference (was
  sharing the trainable decoder LoRA — broke E[KL]=I); banner no longer mislabels it as InfoNCE.
- bio fact_recall: casefold _name_derivable so key-derivable org/work type-words (Wellcome Trust →
  "trust") are dropped from the scored value span.
- qa: question-side un-guessability filter (answer-in-own-question leak in hotpot/triviaqa/multiwoz).
- vanilla encoders take **extra (surprise= kwarg crash); icae dtype cast-back.
- --compile-decoder flag (torch.compile the frozen decoder, dynamic=True).

DOCS:
- REFERENCES.md — authoritative paper/dataset links for every baseline + data source.
- OBJECTIVES.md — the binding objective ladder (MAE-CE → behavioral-KL → provenance-InfoNCE →
  bypass-gap → GRPO), with math + citations.
- graph_thesis.md — the two graph lenses + the literature on making topology load-bearing.
- slotgraph4_design.md (dense edge-state slots, differentiable) + graph_generative_memory.md
  (the "spider web", GRPO-era) design memos.
- DATA_PHASES_PLAN.md: merged the former PHASE_PLAN.md (4 comparison axes, run matrix, invariants).
- Archived the superseded slotgraph/biomem/treemem line + completed reorg plans to docs/history/;
  deleted 9 stale experiment images; refreshed docs/README.md + fixed all broken links.

Remaining before a training run (tracked): H2O eval-ref, titans h-trim, low-severity all-pad edge
guards (gisting/memoryllm), and a behavioral_kl smoke over the per-layer-KV arms.

+++ src/memory/models/icae/encoder.py
@@ -117,7 +117,11 @@ class ICAEBaselineEncoder(nn.Module):
         # over prev-memory + window. base.training=False → HF per-layer ckpt is inert; the trainer's
         # per-window activation-checkpoint (grad_checkpoint_stream) covers this whole forward.
         h = self.base.model(inputs_embeds=inp, attention_mask=attn).last_hidden_state
-        mem = self.norm(h[:, -self.M:, :].float())           # [B, M, d_llama]
+        # Norm in fp32 for stability, then cast back to the running dtype so the memory
+        # carried into next window's cat([prev, token_embeds, slots]) matches (bf16 under
+        # autocast). Without the cast-back, fp32 `prev` collides with bf16 token_embeds in
+        # any eager/no-autocast path — the dtype landmine the debug-real-path convention warns of.
+        mem = self.norm(h[:, -self.M:, :].float()).to(h.dtype)   # [B, M, d_llama]
         return {"mem": mem}, {}
 
     def finalize_memory(self, state) -> tuple[Tensor, dict]:
